Added AllowedCommunitiesComponent
