+++ oarepo_ui/resources/components/base.py
@@ -1,16 +1,10 @@
-from typing import TYPE_CHECKING, Dict
+from typing import Dict, TYPE_CHECKING
 
-from flask import current_app
 from flask_principal import Identity
-from invenio_config.default import ALLOWED_HTML_ATTRS, ALLOWED_HTML_TAGS
-from invenio_i18n.ext import current_i18n
 from invenio_records_resources.services.records.results import RecordItem
-from oarepo_runtime.datastreams.utils import get_file_service_for_record_service
-
-from ..proxies import current_oarepo_ui
 
 if TYPE_CHECKING:
-    from .resource import UIResource
+    from oarepo_ui.resources import UIResource
 
 
 class UIResourceComponent:
@@ -187,103 +181,3 @@ class UIResourceComponent:
         :param ui_links: UI links for the create page, a dictionary of link name -> link url
         :param extra_context: will be passed to the template as the "extra_context" variable
         """
-
-
-class BabelComponent(UIResourceComponent):
-    def form_config(self, *, form_config, **kwargs):
-        conf = current_app.config
-        locales = []
-        for l in current_i18n.get_locales():
-            # Avoid duplicate language entries
-            if l.language in [lang["value"] for lang in locales]:
-                continue
-
-            option = {"value": l.language, "text": l.get_display_name()}
-            locales.append(option)
-
-        form_config.setdefault("current_locale", str(current_i18n.locale))
-        form_config.setdefault("default_locale", conf.get("BABEL_DEFAULT_LOCALE", "en"))
-        form_config.setdefault("locales", locales)
-
-
-class PermissionsComponent(UIResourceComponent):
-    def before_ui_detail(self, *, api_record, extra_context, identity, **kwargs):
-        self.fill_permissions(api_record._record, extra_context, identity)
-
-    def before_ui_edit(self, *, api_record, extra_context, identity, **kwargs):
-        self.fill_permissions(api_record._record, extra_context, identity)
-
-    def before_ui_create(self, *, extra_context, identity, **kwargs):
-        self.fill_permissions(None, extra_context, identity)
-
-    def before_ui_search(self, *, extra_context, identity, search_options, **kwargs):
-        from .resource import RecordsUIResource
-
-        if not isinstance(self.resource, RecordsUIResource):
-            return
-
-        extra_context["permissions"] = {
-            "can_create": self.resource.api_service.check_permission(identity, "create")
-        }
-        # fixes issue with permissions not propagating down to template
-        search_options["overrides"]["permissions"] = extra_context["permissions"]
-
-    def form_config(self, *, form_config, api_record, identity, **kwargs):
-        self.fill_permissions(
-            api_record._record if api_record else None, form_config, identity
-        )
-
-    def get_record_permissions(self, actions, service, identity, record, **kwargs):
-        """Helper for generating (default) record action permissions."""
-        ret = {}
-        for action in actions:
-            try:
-                can_perform = service.check_permission(
-                    identity, action, record=record or {}, **kwargs
-                )
-            except Exception:  # noqa
-                can_perform = False
-            ret[f"can_{action}"] = can_perform
-        return ret
-
-    def fill_permissions(self, record, extra_context, identity, **kwargs):
-        from .resource import RecordsUIResource
-
-        if not isinstance(self.resource, RecordsUIResource):
-            return
-
-        extra_context["permissions"] = self.get_record_permissions(
-            current_oarepo_ui.record_actions,
-            self.resource.api_service,
-            identity,
-            record,
-            **kwargs,
-        )
-
-
-class FilesComponent(UIResourceComponent):
-    def before_ui_edit(self, *, api_record, extra_context, identity, **kwargs):
-        from .resource import RecordsUIResource
-
-        if not isinstance(self.resource, RecordsUIResource):
-            return
-
-        file_service = get_file_service_for_record_service(
-            self.resource.api_service, record=api_record
-        )
-        files = file_service.list_files(identity, api_record["id"])
-        extra_context["files"] = files.to_dict()
-
-    def before_ui_detail(self, **kwargs):
-        self.before_ui_edit(**kwargs)
-
-
-class AllowedHtmlTagsComponent(UIResourceComponent):
-    def form_config(self, *, form_config, **kwargs):
-        form_config["allowedHtmlTags"] = current_app.config.get(
-            "ALLOWED_HTML_TAGS", ALLOWED_HTML_TAGS
-        )
-
-        form_config["allowedHtmlAttrs"] = current_app.config.get(
-            "ALLOWED_HTML_ATTRS", ALLOWED_HTML_ATTRS
-        )
