+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = oarepo-ui
-version = 5.2.4
+version = 5.2.5
 description = UI module for invenio 3.5+
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -33,11 +33,14 @@ exclude =
 [options.extras_require]
 dev =
     oarepo-tools
+workflows =
+    oarepo-workflows
 tests =
     pytest-invenio
     black
     isort
     autoflake
+    oarepo-workflows
 
 [options.entry_points]
 invenio_base.apps =
